requirements.txt
flask
requests
gunicorn
psycopg2-binary
python-dotenv
